Moved Logger Namespace to Marlowe.Logger
