+++ Marlowe/Interpeter.cs
@@ -1,4 +1,5 @@
 ﻿using Marlowe.CSharp;
+using Marlowe.Logger;
 using Marlowe.Utilities;
 using System;
 using System.Collections.Generic;
