+++ Marlowe/Test.cs
@@ -4,28 +4,22 @@ namespace Tester
 {
 	public class SampleClass
 	{
-		private string x = "hello ";
-		private string y = "harry";
-
-		public string z = x + y;
-
+		private static string y = "hello";
+		private int x = 1;
 		public static void Main(string args)
 		{
-			string hello = FindName();
-			string args = "hello";
-			if(x == "hello " || y == "harry"){
-				return "goodbye " + y;
-
-			}
-            
-            {
-				return "hello ";
-            }
+			int z = MyAge();
+			y = "Lewis";
 		}
 
-		public string FindName()
+		public string MyName()
         {
-			return x;
+			return "Lewis";
+        }
+
+		public int MyAge()
+        {
+			return 23;
         }
 	}
 }
