+++ Utilities/Logger/SymbolLogger.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Marlowe.Logger
+{
+    public abstract class SymbolLogger
+    {
+    }
+}
